% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findK.R
\name{findK}
\alias{findK}
\title{findK Function}
\usage{
findK(A, DimSen = 0.97)
}
\arguments{
\item{A}{data matrix A, which represent gene expressions for each cell/sample. rows = genes and cols= cells, first column should be gene names, first row should be cell names.}

\item{DimSen}{is the Dimension Sensitivity, which can be "low", "medium" or "high", default is "medium"}
}
\description{
The main goal of this function is to find a reasonable choice for k.
}
\examples{
findK(A)
findK(A, DimSen="high")
}
\keyword{findK}
