% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Imputation Function}
\usage{
impute(A = 0, k = 0.97, loged = FALSE, W = 0, method = "sNMF")
}
\arguments{
\item{A}{data matrix A, which represent gene expressions for each cell/sample. rows = genes and cols= cells, first column should be gene names, first row should be cell names.}

\item{loged}{should be TRUE if your data is loged, default is FALSE}

\item{W}{is the weight matrix, which is only needed when using WNMF}

\item{method}{could be set to "WNMF" or "sNMF". default is "sNMF"}

\item{K}{is the number of features or estimation of number of sub-cell types, if you have no estimation for k you can initialize it to "low" "medium" or "high" and the findK algorithm will find a good k for you, default is "medium"}
}
\description{
The main goal of this function is to Impute zero values of your data.
}
\examples{
impute(A)
impute(A, method="sNMF" )
}
\keyword{imputation}
