% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{thresholding}
\alias{thresholding}
\title{thresholding Function}
\usage{
thresholding(Raw, Imputed)
}
\arguments{
\item{Raw}{data matrix UnImputed gene expression data, which represent gene expressions for each cell/sample. rows = genes and cols= cells.}

\item{Imputed}{data matrix WNMF/sNMF gene expression data, which represent gene expressions for each cell/sample. rows = genes and cols= cells.}
}
\description{
The main goal of this function is to apply a thresholding step to avoid over imputation
}
\examples{
thresholding( Raw , scaledWNMF)
}
\keyword{imputation}
